+++ addons/keh_network/network.gd
@@ -642,14 +642,20 @@ func _on_snapshot_finished(snap: NetSnapshot) -> void:
 	snap.input_sig = player_data.local_player.get_last_input_signature()
 	snapshot_data._history.push_back(snap)
 	
+	var popped: int = 0
+	
 	# Ensure the snapshot container remains with a reasonable amount of data. If this loop
 	# occurs on client, either:
 	# 1 - there was a massive data loss
 	# 2 - input data was not polled
 	while (snapshot_data._history.size() > _max_history_size):
 		snapshot_data._history.pop_front()
+		popped += 1
 	
 	if (!has_authority()):
+		# On clients update the prediction count for each entity.
+		snapshot_data._update_prediction_count(1 - popped)
+		# Also, there is nothing else for clients to do here, so bail
 		return
 	
 	# Iterate through remote players and update as required